feat(歌曲): :sparkles: 搭建起来歌曲仓库基础

+++ apps/music-api/music-core/music-infrastructure/music-infrastructure-persistence/src/main/java/io/gitee/pkmer/core/infrastructure/persistence/song/mybatis/Song.java
@@ -19,7 +19,8 @@ public class Song {
     @SnowflakeId
     private Long id;
 
-    private Integer singerId;
+
+    private Long singerId;
 
 
     private String name;
